Added ObjectProxy for 'active' persistable objects
The (persistable) ObjectProxy just keeps the UUID of the object you want to persist
and will try and find the object in the loop when it gets unbundled, failing if it
is not found.  This is a convenient way to 'persist' a proxy for an object that isn't
itself persistable.  Of course, this object must then be inserted into the loop before
the proxy is unbundled so it finds it.